Adição de view e controller para cadastro

+++ src/main/java/com/atividade/SpringDataEntity/repository/EspecialidadeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.atividade.SpringDataEntity.model.Especialidade;
 
-public interface EspecialidadeRepository extends JpaRepository<Especialidade, Integer> {
+public interface EspecialidadeRepository extends JpaRepository<Especialidade, Long> {
     
 }
